rename and show translations name
