+++ src/components/Navbar/SettingsDrawer/ReciterSelectionBody.module.scss
@@ -1,0 +1,9 @@
+.reciter {
+  margin-top: var(--spacing-small);
+  display: flex;
+  align-items: center;
+  font-size: var(--font-size-large);
+  & > input {
+    margin-right: var(--spacing-small);
+  }
+}
