--- src/components/Navbar/SettingsDrawer/SettingsReciter.module.scss
@@ -1,9 +1,0 @@
-.reciter {
-    margin-top: var(--spacing-small);
-    display: flex;
-    align-items: center;
-    font-size: var(--font-size-large);
-    & > input {
-        margin-right: var(--spacing-small);
-    }
-}
